% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular_plot.R
\name{triangular_plot}
\alias{triangular_plot}
\title{Triangular Plot}
\usage{
triangular_plot(df, varx, vary)
}
\arguments{
\item{df}{a dataframe}

\item{varx}{a string name for variable x in the dataframe df}

\item{vary}{a string name for variable y in the dataframe df}
}
\value{
A plot with diamond markers of \code{vary} versus \code{varx}
}
\description{
This function produces a ggplot plot of \code{vary} versus \code{varx} with
triangular markers
}
\examples{

df_test <- data.frame(x = rnorm(10), y = rnorm(10))
triangular_plot(df_test)



}
